MVVM-ify DFOToolbox with Prism.
Move code from MainWindow class to a MainWindowViewModel class.

Use DelegateCommands from Prism, with view-layer commands calling to
viewmodel-layer commands when applicable. Use a XXXCommandCanExecute
property to communicate to the view if menu options should be enabled,
etc, and use INotifyPropertyChanged to notify of changes to whether
the command can execute.