Fix bug where the Save menu item was sometimes not enabled.
Don't set a delegate on commands indicating if the command can execute.

WPF has strange ideas of when it should requery commands to see if they
can execute. And when it does, it supposedly requeries ALL commands.

Prism's DelegateCommand's RaiseCanExecuteChanged() may or may not work.
I think I tried it before and it did not work.
But even if it does, WPF's implementation of command checking seems to
requery all commands which could get inefficient.

We already have our own plumbing in place for enabling/disabling UI,
we just need to totally remove all traces of WPF's lackluster method.